Update fruit store context table definitions
Remove fruit repository
Update fruit routing, model and controller

+++ src/main/kotlin/ktorwebapi/models/Fruit.kt
@@ -1,3 +1,3 @@
 package ktorwebapi.models
 
-data class Fruit(var id: Int, var no: String, var description: String)
+data class Fruit(var id: Long, var no: String, var description: String)
